Set up files for user interface
There's a basic index control page set up but it doesn't do anything
yet (as I have not set up the api for it to connect to). Next steps:
convert everything to blueprints, build the api (properly) add javascript

Loading things with the testing enviroment and big files such as bootstrap
and jquery is really annoying however.