Stacks are now working !!!!
This took a while and probably requires a little bit more testing
but it works now. I've also added global variables for both fadetime
and bpm. I say global but currently they only apply to the stacks. This
allows the user to have more control over the stacks without having to
restart it (do note however that they will be able to change without losing
their place in the stack)