Added script in UI to send the scenes to the api
I had a few issues as I left out a + when concentrating the url
strings and the debugger was really unhelpful. Next steps are
to implement a fader bar, and the bpm change for the stacks.

However I plan to convert the db to sqlite first, and then start keeping
track of those variables there.