Added BPM & fade time to ui and DB
The bpm and the fade time are now a part of the database so
the values are kept session to session, I also implemented them in mainui
and they send both as their own things and with the channel/scene/chase
change. I should properly implement fadetime selector as something
logarithmic however but not today.